Test Build of Tracker with WMA Music Support
This test build of the tracker can now play the original WMA music files. Just copy the game's WMA music files into the tracker's sound folder and it will play them during the game. You can replace these files to customize them. You can also use the MP3 and WAV formats if you generate a complete set of files in those formats. That means you need every single song in mp3 format or wav format. The priority the soundtrack sets will be searched for are MP3, WAV, WMA. So if there is no MP3 set, it will look for a WAV set, if the WAV set isn't there then it will look for the WMA set. If none of the 3 sets are found no music will play.